Fixing a potential crash in the driver.
I don't know if this will actually fix it because it basically never happens so I can't easily verify.
In the syslog, I see:

[ 2836.058498] BUG: unable to handle kernel paging request at ffff8800e8d7f000
[ 2836.058519] IP: [<ffffffffa01e5a26>] f_proc_startupdate+0x96/0x36c [ppm]
[ 2836.058534] PGD 160c067 PUD e57067 PMD f9e067 PTE 80100000e8d7f065
[ 2836.058545] Oops: 0003 [#1] SMP
[ 2836.058553] CPU 0
[ 2836.058555] Modules linked in: ppm(O) ipv6 binfmt_misc scsi_mod acpiphp pci_hotplug crc32c_intel ghash_clmulni_intel microcode ext4 mbcache jbd2 crc16 aesni_intel cryptd aes_x86_64 aes_generic dm_mirror dm_region_hash dm_log dm_mod [last unloaded: scsi_wait_scan]
[ 2836.058600]
[ 2836.058606] Pid: 6669, comm: java Tainted: G           O 3.4.62-53.42.amzn1.x86_64 #1
[ 2836.058616] RIP: e030:[<ffffffffa01e5a26>]  [<ffffffffa01e5a26>] f_proc_startupdate+0x96/0x36c [ppm]

And since ffff8800e8d7f000 is page-aligned (last 12 bits set to 0), my guess is that the java process (which has a long command line of 20kb) arguments get truncated to 4096, and by setting null in [args_len] we might potentially write in the first byte of the next page, causing a fault under extremely rare conditions.

Of course this is all just my mental masturbation, we'll see if it happens again