Use worker instead of nginx master, issue #72
Happens that the nginx master is not reported by sysdig agent
because its cpu/memory/activity is very low compared to the workers
so the whole nginx metrics are excluded.

Matching the nginx worker should do the trick, all the workes are
grouped in a single program so it should not mess up anything.