Tests for falco functionality in agent.
Add unit tests for falco in agent. This tests up to the level of
interactions between the inspector, agent, and logging module which
receives user events. Other functionality is not tested (but also wasn't
new to these changes).

The test fixture and test cases are in falco_events.cpp. The following
tests are covered:

 - Matching a rule in the the main falco_rules.yaml file.
 - Matching a rule in the user_falco_rules.ymal file.
 - Matching nothing (i.e. an event stream that should not match).
 - Disabling the main rule via direct reference and wildcard.
 - Disabling the user rule via direct reference and wildcard.
 - Disabling falco entirely, does not match a set of events that
   previously matched.

To provide the streams of events, use 3 small (10k) trace files. These
and the corresponding .yaml files are saved in the resources/falco
subdirectory.