Unit test for clone proc fix.
Unit test that verifies fix for
https://github.com/draios/sysdig/issues/664.

The test generally does the following:
 - Ensures that a stale process exists
 - Starts another process with the same pid as the stale process, in a pid
   namespace (which counts as "in a container").
 - Checks to see if the stale process information is used.

To distinguish between the stale process and up-to-date process, use the
working directory of the process. The stale process sets its working
directory to "/dev".

Prior to the fix for 664, the stale process would be used and the
working directory for the new process would be /dev. With the fix, the
stale process information is detected and removed and the working
directory for the new process is the working directory of the
test (i.e. not /dev)

The details are a bit more complicated:

 - clone_helper() is a general utility function to create a stack, call
   clone() to start a thread, and (usually) wait for it to
   complete. This is used to to start all the other actions.
 - stop_sinsp_and_exit() creates the stale process. It chdir()s to
   "/dev", stops the inspector, and returns.
 - do_nothing() simply immediately returns.
 - launcher() simply launches a do_nothing() thread,
   synchronized with the main test.
 - A condition variable passes control between the main test thread and
   launcher() and back.

The actions performed by the main test are:

1. Start the launcher() thread. It waits to be signaled by the
main test thread.
2. Start a stop_sinsp_and_exit() thread and wait for it to finish. The
pid used by stop_sinsp_and_exit() (aka recycled_pid) will be used again.
3. Write the recycled pid to /proc/sys/kernel/ns_last_pid, which ensures
that the next process created will have recycled_pid.
4. Signal launcher() to start a thread.
5. launcher() starts a do_nothing thread, waits for it to
complete, and signals the main test thread.

Note that there's a race condition between steps 3-4. It's possible that
some completely unrelated process started before launcher()
started a do_nothing() thread. However, as long as that process does not
have a working directory of /dev/ it will be possible to distinguish it
from the stale process.

To verify the actions, the filter selects all events related to pid
recycled_pid. It should see:
 - a clone()
 - several events with pid=recycled_pid and cwd=<where the test is run>
 - after stop_sinp_and_exit chdir()s to /dev, several events with
   pid=recycled_pid and cwd=/dev
 - a second clone()
 - events with pid=recycled_pid (the do_nothing started by
 launcher) and cwd=<where the test is run>

If after the second clone any event with pid=recycled_pid has a cwd of
/dev/, the test fails.