nginx config to add ssl term/auth to dcos
A sample nginx config that adds ssl termination and emulates
authentication token generation. This may be useful for automated tests
as well as when testing against non-enterprise DC/OS.

It listens on 5050/5051 + ssl, proxying all requests to a downstream
server and requiring an authentication token. The specific endpoints /
and /version do not require authentication.

It also listens on 443, requiring a hardcoded username/password and
returning an authentication token.