App check changes to support templated urls.
This is the app check half of the changes to allow the mesos url/auth
token to be a templated value which is filled in by the agent.

When running the app check, a top-level json attribute 'conf_vals'
contains templated values, and are expanded in the python sdcheck like
the machine-level information like port are expanded.

In the case of mesos-master/mesos-slave, the templated values are
'mesos_url', which is replaced with http/https as appropriate + the
autodetected or configured hostname, and 'auth_token', which is
generated/refreshed in the agent and passed to the app check.

Make the regex looking for template patterns non-greedy so you can
specify multiple templated values in one string.

Finally, update sdchecks.py, which can be used to run app checks
individually from the command line, to take an argument containing a
dict of conf_vals which is used to resolve templates.