Agent changes to support configurable url/auth
A new class app_process_conf_vals, associated with the app_process
class, is responsible for holding any non-system additional
configuration that should be passed to an app check. It has a vals()
method that is used as the value for a 'conf_vals' attribute that is in
the json passed to an app check.

app_process_conf_vals must be subclassed. The only current subclass is
mesos_conf_vals, which takes the configured dcos credentials and the
mesos state uri discovered/configured in detect_mesos() and passes it to
the app check. It contains a mesos_auth object, which is responsible for
generating and refreshing the auth token every time the app check is called.

A new function match_checks_list is responsible for looping over the set
of app checks, crearing app_process objects as necessary, and in the
case of mesos-master/mesos-slave/mesos-agent ceeaing a mesos_conf_vals
to do the json marshalling.

detect_mesos/detect_local_server now work with a configurable port,
which is either 5050 when looking for a mesos-master or 5051 when
looking for a mesos agent/slave. When a server is discovered on port
5051, the main mesos fetches for marathon, etc. are all skipped.

The work of determining when to refresh the mesos token is now handled
by the mesos_auth class, so it's a bit simpler within the agent.