Pass mesos/marathon basic http creds to app check
When mesos/marathon credentials are defined, pass them to the
messos_master/mesos_slave/marathon app checks. The credentials are only
used if dcos authentication is not specified, resulting in no auth
token.

The credentials are passed to the app check via the templated value
mesos_creds/marathon_creds, and is filled in by the
mesos_conf_vals/marathon_conf_vals objects when looping over the app
checks.

Also remove some apparently unused code for the marathon app check that
allowed providing a user/password. I don't believe this was used and is
deprecated by passing the creds in marathon_creds.