Fixes for stale thread handling.
With the changes in
https://github.com/draios/sysdig/commit/6b6ffb697e5146989ec7f0a0f3ba99ccc9c7118e,
stale thread detection is now time-sensitive, and a thread state must be
at least 2 seconds old to be considered stale.

To address that, have the first process whose pid will be recycled and
whose state is not cleanly removed wait 5 seconds, so it meets the test
of being at least 2 seconds old.