Add security capabilities to agent.
Add the ability to load a set of security policies provided by the
backend, match events against those policies, and take actions based on
those policies.

The overall design is fairly generic. In this version, policies are tied
to falco rules. Each policy refers to a set of falco rules that are
provided in a corresponding falco ruleset, either by name regex or list
of tags.

(Note that all of these capabilities are separate from the existing falco
engine that is embedded in the agent, which remains untouched).

A policy consists of:
  - a scope (collection of key=value tags)
  - a rule filter (glob on rule name or tag)
  - a set of actions (can either be stop container, pause container, or
    perform capture)
  - A flag apply_scop
  - enabled flag.

security_mgr.cpp is the top-level object managing the set of security
policies and is responsible for receiving messages from the
backend (e.g. the list of policies) as well as sending messages to the
backend (e.g. policy events when policies trigger). security_policy.cpp
contains the object for an abstract security policy class as well as a
falco policy subclass, and is responsible for performing
actions (stop/pause container) when a policy triggers.

security_policy.cpp uses cointerface/coclient to stop/pause containers,
and uses the existing capture job handler mechanism to start sysdig
captures. All actions are performed in the background. A policy event
message is only emitted once those actions are completed (the container
is paused/stopped, or the capture job request has been submitted).

The following new top-level messages are sent to/from the backend:

 - POLICIES: sent from the backend to the agent, contains the set of
   policies to use.
 - POLICY_EVENTS: sent from the agent to the backend, sent whenever a
   policy is triggered. These are generally batched together and sent
   every 10 seconds (configurable), but can be sent immediately when an
   action is a capture action, to make sure the policy event message is
   received before any capture chunk messages.
 - THROTTLED_POLICY_EVENTS: to avoid floods of POLICY_EVENT messages, a
   token bucket limits the rate of POLICY_EVENT messages. When the rate
   is exceeded, the events spill over into a THROTTLED_POLICY_EVENT
   message that simply maintains a count rather than full details on the
   event, and skips all actions.

The following new config options control the behavior of the security
capabilities:

 - security: enabled: <bool>: determines whether to create the security
   manager object.
 - (testing-only) security: policies_file: <path>: points to a messages
   protobuf message as text.
 - security: report_interval: <ns>: how often to report batched
   POLICY_EVENTS mesages. Default 1 second.
 - security: throttled_report_interval: <ns>: how often to report
   THROTTLED_POLICY_EVENTS messages. Default 10 seconds.
 - security: policy_events_rate: <rate> and security:
   policy_events_max_burst: <count>. Controls the rate at which
   policy events can be generated before spilling over into throttled
   policy events counts.
 - security: send_monitor_events: <bool>. Whether or not to send sysdig
   monitor user events for each policy event. (Default false).
 - sysdig capture: compression level: <int>. Allows changing the
   compression level for sysdig capture chunks.

There are also several changes to handle potential bursts of policy
events. These changes add throttling to prevent bursts of events from
making it to the backend as well as efficiency improvements to better
handle bursts of events.

 - Instead of using the protocol_queue for policy event messages, use a
   new data structure synchronized_policy_events, which is basically a
   mutex-protected draiosproto::policy_events message. put() appends a
   new set of policy events to the array. There's still a maximum number
   of events MAX_QUEUED_POLICY_EVENTS, which causes put() to return
   false.
 - connection_manager's main loop now reads from
   synchronized_policy_events in each loop. When any messages are found,
   they are sent to the collector. Together, these changes allow bursts
   of policy events to be grouped as much as possible.
 - policy events are now throttled by a token bucket that enforces a
   burst/rate for a given (policy_id,container_id) group. Any time a
   policy event is throttled by the token bucket, a separate message
   throttled_policy_event is used instead to increment a
   count. Throttled policy event messages are sent every 10
   seconds (default) and aggregate counts, but not policy event details,
   across all containers/policies. throttled_policy_event messages are
   sent via the protocol_buffer queue used for metrics/capture messages.
 - Centralize all of this in security_mgr::accept_policy_event(). This
   changes some of the other security policy functions that used to
   append to a list of policy events, and generally simplifies things.
 - In security_mgr::process_event, use run_on_interval classes to
   schedule the periodic tasks of checking for completed actions,
   sending policy events messages, and sending throttled policy events
   messages.

Now that there's a lot of additional work within sinsp_worker, start
measuring the delay from the time a message is queued for processing by
the connection manager and the time the message is actually sent to the
collector. The main change is to change protocol_queue_item from a
simple string to a struct containing the string, message type, and time
at which the message was enqueued. In
connection_manager::transmit_buffer, measure the elapsed time and if
it's over 5 seconds log a warning with the message type and delay.

Finally, reduce one source of delays by changing the priority of policy
event messages from MEDIUM to LOW. That way, a burst of policy event
messages won't block metrics (and more importantly, commands).

Also make some initial improvements to sending sysdig
captures. Previously, the connection manager loop only had an
opportunity to send a message to the collector every 200ms, even when
there were pending messages to send to the collector. Every connection
manager loop would wait 200ms for a message from the collector (openssl
poll with a 100ms timeout + openssl read with a 100ms timeout). This
artifically slowed down messages from the agent, including metrics and
policy events messages. Make several changes to fix this:

 - Instead of doing a 200ms blocking read waiting for data from the
   collector, do a non-blocking poll and only try to read a message if
   the poll shows pending data. This eliminates the 200ms delay per
   loop.
 - To ensure the connection manager loop doesn't spin, increase the
   timeout for the queue read from 100ms to 300ms. This way, when there
   is nothing to send or receive the connection manager loop will still
   wait 300ms.
 - Add a token bucket that throttles all traffic used by all sysdig
   captures. It's controlled by the config option "sysdig_captures_rate",
   with a default of 1MB/sec.
 - When throttling occurs, the set of jobs is rotated so each job gets
   an somewhat equal change to send data.
 - Check for pending jobs every 100ms instead of every second. This
   avoids burstiness in captures as the token bucket reclaims bandwidth.