Explicitly disable libunwind in gperftools.
On my system, I need to have libunwind installed (because of perf), and gperftools seems to automatically pick it up during the build, and then the linking of dragent fails because we don't link against libunwind.

I haven't done a lot of research but that seems a plausible explanation, and the fix in the commit seems to work for me.