Unit tests for thread state traversal.
Add unit tests for thread state traversal. There are tests for:
 - A simple chain of processes with parents and no loop.
 - A single process with no parent.
 - A chain of processes with a loop.
 - A lollipop with a loop starting partway up the heirarchy.

The traversal actually occurs in a separate thread with a promised
result that detects whether or not the traversal finishes.

In each case, we verify which nodes are visited and whether a loop was
detected.

There's also a test that intentionally waits for a long time to verify
that the future result querying works as expected.