Get stale thread test working
It turns out this test was succeeding by accident. Adding the
EXPECT_EQ(clones_seen, 1); test was a step in the right direction,
although the test needs to occur after the main actions of setting up
the inspector and running through the tests.

The problem was that evt->m_tinfo is set to NULL as a part of the stale
thread cleanup, meaning that the event itself didn't have a tightly
linked threadinfo. This meant that the filter in the filter lambda
wasn't selecting the critical clone events for the recycled thread, so
it would never get to the cwd check:

EXPECT_STRNE(tinfo->get_cwd().c_str(), "/dev/");

If I recall, I used to check recycle_pid not in the filter but in the
callback, so it would end up seeing some (non-clone) events where it
could check the cwd.

I'm not sure if it's a bug that evt->m_tinfo is NULL when there actually
is a thread associated with the event, though. If so, that requires a
sysdig change.

evt->get_thread_info(), however, does return accurate thread
information, so use that instead to get the thread info.

Once you do that, the callback logic becomes much simpler--the *only*
events it will see are ones releated to the second thread using the
recycled pid, so you can always check the cwd and verify that it is not
"/dev".