Fix race condition in arguments to thread.
There's a possible race condition with the captured variables for the
runner thread (specifically the promise object), which are references to
variables on the stack. When running under valgrind i noticed that the
runner thread was accessing already-deleted memory.

Fix this by switching from captured references to arguments to the
runner function. Preserve the promise by using std::move. The others are
ok as they are pointers/copied.