Add golang/grpc to build.
In the docker builder, also install the golang package. This happens to
be go 1.7. I've also tried with 1.6 and it builds properly.

Add grpc to bootstrap-agent. The normal build process for grpc involves
git clone + git submodule update --init, which pulls in a lot of
additional dependencies that we don't need. Instead, we download our
own repackaged tarball that has only grpc and the necessary third party
libraries (protobuf 3.0.0, boringssl) to build the targets we
need (grpc_cpp_plugin, static_cxx, and static_c). It also symlinks with
the dependency zlib instead of grpc's own copy.