Adding go-based interface program to agent
Add new messages/grpc interfaces to sdc_internal.proto that handle the
messages/rpc calls we'll use to the cointerface process.

The go code is under userspace/cointerface. Below this directory, files
are organized as if it were a GOPATH (project root). During the build
process, additional required packages will be installed below this
directory as a result of go get. The interesting subdirectories are:

 - cointerface: contains the cointerface program
 - coclient: a test client program that can send messages to cointerface.
 - draiosproto: contains autogenerated protobuf code from draios.proto.
 - sdc_internal: contains autogenerated protobuf code from sdc_internal.proto.

Once built, the server and client programs will be built at
userspace/cointerface/{cointerface,coclient}. "go fmt" is also run
during the build to ensure that the code is formatted following go's
guidelines.

Additionally, during packaging cointerface will be copied to
/opt/draios/bin/cointerface.  There are new cmake targets
userspace/cointerface/sdc_internal_pb_go and
userspace/libsanalyzer/sdc_internal_pb_cpp that run protoc with the
necessary arguments to generate the go/cpp code from the .proto
files. Also note that when generating code for sdc_internal, it has to
alter the import namespace for draios.proto to be its own package.

On the go side, we use gogo/protobuf/proto to generate protobuf
code. Evidently it's a better go generator for protobufs, with faster
marshaling/unmarshaling code. It's wire-compatible with the existing
format.

cointerface's log output is json so it can be easily parsed by the
agent. Unfortunately, the default seelog json output doesn't escape
quotes, so add a custom formatter that takes the pid/level/message as a
struct, serializes it, and writes it as json.

cointerface uses a domain socket at /opt/draios/run/cointerface.sock for
grpc communication, and supports the following rpcs:
 - ping. This takes a (numeric) token and returns that token + the
   current pid and memory usage in kb. This is used as an example to
   show how the grpc code works. It's also used by the subprocess
   watchdog to check liveness/memory usage.
 - run docker command. This will be used by project anna to manage
   containers as a result of policy actions.

To perform the docker command, we use the go docker client. We request
api version 1.18, which is the oldest documented version on
https://docs.docker.com/engine/api/ and corresponds to docker
1.6. Should definitely be old enough to have a wide degree of
compatibility. When running inside a container, prepend the value of
SYSDIG_HOST_ROOT to the docker socket path.