Add c++/agent support for cointerface
Add c++ side of grpc connections to the cointerface process.

sdc_internal.pb.{cc,h} has been regenerated to include the definitions
from the updated sdc_internal.proto, and sdc_internal.grpc.{cc,h} has
the grpc code. The grpc code was actually generated using protoc 2.5,
but with a grpc 3.0.0 plugin. This appears to be compatible with 2.5,
generating code that compiles and links alongside protobuf 2.5
libraries. We'll need to double check the format of any protobufs
generated within cointerface to see if they can be directly sent back to
the collector.

Add a make target sdc_internal_pb_cpp to generate c++ code from the
.proto files. Like the go target, it's not associated with ALL so it
will only run when specified.

coclient.{cpp,h} is a wrapper around the grpc code that provides async
access to the cointerface process. The only main interface coclient has
is ping(), which takes a token and callback function. The callback is
called when the response has been received. next() is used to poll for
responses.

Most of the details of setting up the rpc are in prepare(), and are
intended to be general purpose for any rpc of the form response_msg =
rpc_func(request_msg). Hopefully it will also work for streaming
responses, but I haven't tried that yet.