Unit tests for memory dumps
Add unit tests for memory (and standard) dumps.

The test setup includes a sinsp_worker. You schedule dump requests for
the sinsp_worker using sinsp_worker::queue_job_request and receive the
results of the requests using the shared protocol_queue used by the
sinsp_worker.

The tests open specific files intermingled with the dump
requests. Afterward, the resulting trace files are read and verify that
the specific file opens are seen.

There are support methods that allow requesting a dump with a generic
tag, waiting for all dumps to complete for a given set of tags, and
reading trace files looking for a set of tags.

Test cases include:

 - A simple standard dump with no memory dump (back in time) capability.
 - Using memdump to capture a file open before the dump request.
 - A test for overlapping dump requests. In the test's worker thread,
   write to a series of files every second and schedule a series of dump
   requests every second. Each dump request is for 1.5s before and 1.5s
   after, meaning that each dump will capture the prior file open, the
   current file open, and the next file open.
 - A test for hitting the cap on the number of outstanding captures
 - A test for filling the queue of dump requests before the sinsp_worker
   can dequeue them.

The last 2 tests both work by triggering 11 captures. The first 10
should succeed and the 11th should fail.