Add unit tests for coclient.
Add unit tests for coclient. The test starts a cointerface process using
a custom unix socket. Cointerface is run with -use_json=false. The test
captures its plain text logs and logs the messages at debug level.

Test coclient's ping() method to ping the cointerface process. The
callback ensures success and that the ping token is returned in the
pong.

Also test the PAUSE and STOP docker commands. The test starts a simple
busybox container and then either pauses or stops it. In each case,
we're looking for a successful rpc and a successful result from
cointerface + no error string.