Move capture job handling to separate thread
Move the work of managing and sending capture jobs to a separate thread
capture_job_handler:

Move the code that used to be in sinsp_worker related to
starting/stopping capture jobs to capture_job_handler.{cpp,h}. The most
important methods are:

 - run(): run forever, doing the work of processing job requests and
   sending capture contents to the connection manager
 - process_event(): called from the sinsp_worker's thread, update any
   active capture jobs with a new event
 - queue_job_request(): called from the sinsp_worker's thread, pass a
   capture job request on to the capture job handler thread. This can
   now return immediately if it's clear the job request will fail.

sinsp_worker actually still has a job request queue m_dump_job_requests,
but this is now only used by the connection manager to pass along sysdig
monitor JOB_REQUEST protobuf messages to the sinsp_worker. These go
through the sinsp_worker so it can create necessary sinsp_dumper objects
associated with the sinsp_worker's inspector.

The memdumper moves from the analyzer to the capture_job_handler.

A lot of the code related to actually starting/stopping/processing
events for jobs has been moved into dump_job_state (renamed to
capture_job), which now is more of an object than a struct.

There used to be a pretty separate distinction between standard and
memdump jobs. Now they're held in a single list and via a single state
object. The only difference is their m_past_duration_ns values differ
and that when created, memdump jobs use the memdumper object to dump the
memory state to disk, filter it to get matching events, and return a
sinsp_dumper associated with the filtered file.

The shared objects it uses to communicate with other threads are:
 - A protocol_queue (m_queue) to send messages to the connection
 manager.
 - a queue of requests (m_dump_job_requests). This is all within
 queue_job_request() but is the mechanism to get job requests from the
 sinsp_worker thread to the capture job handler thread.
 - an atomic bool enable_autodrop. This is set by the capture job
 handler thread and read by sinsp_worker. When changed, it tells the
 sinsp_worker to enable/disable autodrop. Note that autodrop is now
 disabled when there are any capture jobs, as compared to only standard jobs.

The synchronization primitives used to coordiate the two threads are:

 - capture_job_handler::m_jobs_lock: a Poco::RWLock that protects access
   to the list of jobs. process_event() acquires a read lock and
 - memdumper::m_state_mtx: A mutex that controls access to memdumper's
   in-memory buffers. process_event() and add_job() are synchronized.
 - capture_job::m_mtx. A mutex that controls access to a job
   state. process_event() and stop() are synchronized.

Capture job requests are stored in capture_job_handler::dump_job_request
structs. They are pretty much unchanged from sinsp_worker. One addition
is m_start_ns, which allows specifying an exact start time for memdump
jobs, and m_dumper, which is a sinsp_dumper object created from the
sinsp_worker thread, using its inspector. If not set, it either uses the
last event time or worst case the current time.

The capture_job_handler doesn't have its own inspector--it uses
sinsp_worker's inspector. It only uses it for methods called by
sinsp_worker (process_event). One slight cheat is that when creating
jobs, the filter is associated with the sinsp_worker's inspector. That
filter/associated inspector is only used when matching events in
process_event(), though.

In the security manager, send job requests to the capture job handler
instead of sinsp_worker and set the job start time to be the time of the
associated policy event. Also, if an attempt to queue a job request
fails (like hitting the max # outstanding captures or the queue being
full), fail the policy action instead of lying about its success. For
policy events that have failed capture actions, send them via the normal
queued process instead of sending them immediately.

Rearrange the unit tests to get them working as well. Instead of running
sinsp_worker in the foreground, run it (and the capture job handler) in
the background using Poco::ThreadPool. The test code moves to the
foreground, and there are small changes to packet sizes, cleanup order,
etc. Remove the test that checked for a full queue--it was too
unreliable as the capture job handler really could empty the queue at
any time.

One slightly unrelated small change--Improve logging for throttled
security events by adding 'sec_evts' to the log message and add a
overall count of the number of throttled events.

Finally, remove the remnants of support for completely dropping events
until autodrop had kicked in. This was originally added in
https://github.com/draios/agent/commit/511675b3b, but we might as well
include whatever events we can in the capture.

There's now a fair amount of dead code within memdumper--it used to
manage a set of memdump jobs, but all of that is now handled by the
capture job handler. Not sure if we want to remove that right away or
not.