Disable autodrop differently on captures
Instead of disabling autodrop when a sysdig capture starts, set the
sampling ratio to 1. This ensures that the driver acks the sampling
ratio, causing the analyzer loop to start flushing metrics again.

Also pin the sampling ratio by setting a (new) variable
capture_in_progress. If true, tune_drop_mode won't be called.

Finally, try to clean up the use of m_new_sampling_ratio. Now it's only
set in start_dropping_mode() and reflects the desired dropping
ratio, which is finally set either when the DROP_X event is received or
after a timeout.