Keep and log metrics on security policy results
Keep track of the number of events processed and what eventually
happened for each event for each policy. This is handled by the class
security_policy::evt_metrics, which has an array of counts.

In security_policy::process_event, increment these counts and reorganize
the flow to separate all the possible exit paths.

A new run_on_interval controls when to log these counts (default 60 seconds).