Skip agent events captures (#585)
* Skip all agent events in captures when configured

There's been an option capture_dragent_events, default false, for a
while. When false, in sysdig captures any activity for the dragent
process doing the capture was excluded.

Since that option was added, we have a larger collection of dragent
processes, however (sdchecks, cointerface, etc). It'd be nice to exclude
their activity as well.

One easy way to handle this is to rely on their session id, which is the
same for all the dragent processes as long as you do getsid() after the
agent has backgrounded itself.

So in both memdumper and the capture job handler, get the session id and
exclude events for that session from captures.

Also in the capture job handler, move the skipping up so it's done once
for all jobs instead of for each job.

* Also excude switch events.

https://github.com/draios/agent/commit/caa87f849450ccc11303c4b75ff00334cc20102a
made sure switch events always ended up in captures, because at the time
it was necessary to properly capture cpu usage.

We capture cpu usage via ioctls now, so we can exclude these events like
other events.

* Unit test for excluding dragent events.

Unit test for excluding dragent events from captures. Verifies that for
a simple capture with a back-in-time start that the resulting trace file
has nothing for the program "tests".

This will pass once https://github.com/draios/sysdig/pull/906 is merged.