SMAGENT-248: Internal metrics (#441)
* Send internal agent metrics

Add the ability to report internal metrics from the agent. These metrics
include:

 - The number of processes, threads, containers, java processes, and app
   checks.
 - The cpu usage/memory usage of all the agent-related subprocesses.
 - Whether or not mesos delegation is enabled and whether a mesos
   environment has been detected.
 - The percentage of cpu time spent in flush() (the fp field of the
   periodic metrics log message).
 - The time spent in flush() (the fl field of the periodic metrics log
   message).
 - The number of sysdig events processed and dropped.
 - The current sampling ratio.
 - The number of log messages at every log level.

We'll add k8s internal metrics later once the
cointerface/k8s/orchestrators work is complete.

internal_metrics.{cpp,h} is the class managing the internal metrics. It
has methods to set/get all the above and to fill in a protobuf with the
values. The internal metrics object is created in the top dragent and
passed to the sinsp_worker/analyzer.

Modify the logger to take the internal metrics object and notify it for
every log message. The internal metrics object bumps counts for each
message.

The internal metrics are a new message "internal_metrics" that is a part
of the metrics message.

There's a fair amount of changes to allow monitoring the cpu/memory
usage for the agent subprocesses:

 - The main agent process knows the set of subprocess pids, from
   the subprocess logger, so it has to pass that down to the
   internal_metrics object, as a map name -> pid. That's protected by a
   read-write lock.

 - Add a name to each subprocess watchdog state so the cpu/memory usage for
   each subprocess can be tagged with that name.

 - Fix the subprocess state to also track the main agent's pid--it
   doesn't read it from the log files like other subprocesses.

 - internal_metrics::update_subprocess_metrics, called from the
   analyzer, updates the cpu/memory usage based on the provided map of
   subprocess names/pids.

* Make shared flags atomic

Several shared flags to signal shutdown, etc. between threads were
flagged as volatile. This isn't really thread-safe.

Make them atomic instead, which is thread safe.

* Whitespace diffs.

Committing separate from other changes.

* Add new files to dragent.project.

* Misc minor fixes.

Explicitly type some variables.

Be consistent about void arguments to lambdas when starting
subprocesses.

Allow implicit conversion of watchdog_state/pipe_manager.

Reorder some log messages related to signaling.

Minor typo fixes.

* Add unit tests.

new unit test uses the internal metrics object, sets values for all the
various metrics, verifies that the get_XXX methods work and that the
protobuf message has the expected format.

* In policy events, flag dropped sinsp events

Add a new boolean field sinsp_events_dropped to the policy_events
message. If true, it indicates that inspector events were dropped around
the time the policy event was created.

Connect the security manager and analyzer and use a method from the
analyzer recent_sinsp_events_dropped() to fill in the field for the
policy_events message. It returns true if the sum of n_drops and
n_drops_buffer is > 0.

* Disable one memdump unit test.

It's not reliably passing. Will fix in a separate PR.