Move push_notification() into capture_job_handler
Move push_notification(), which injects an artificial event into the
event stream that reflects the start of a capture job (or later, other
container/k8s events) from the memdumper to the capture job handler.

That way, the event will also be written to any capture jobs that are in
progress.

Also fix the date of the test event emitted by the logger to the current
time so it will be included in capture files.