Disable memdump nomem SMAGENT-582 (#607)
* Ensure memdump size is available at startup.

At memdump startup, create a temporary shared memory region with the
expected memdump size. If this fails, log an error and disable memdump.

This ensures that if the agent was started without being configured with
the necessary shared memory size (i.e. --shm-size=350m via docker) that
it will still run, albeit without memdump.

* Skip memdump captures when memdump disables itself

With the prior commit, the memdumper will be disabled if isn't enough
shared memory available for the memory buffer. In addition, the
memdumper can disable itself if it has a problem opening the shared
memory regions or switching states.

In the capture job handler, check both the configuration as well as the
memdumper status before scheduling a back-in-time capture. If either is
disabled, fail the capture job saying that back in time capture is not
enabled.

* Bumping protocol to 4 (safe memdump handling)

Bumping protocol to 4 to identify agents that can safely have memdump
enabled despite not having enough shared memory.