Force capture job handler shutdown during tests. (#610)
ThreadPool::defaultPool().stopAll() only waits 10 seconds for each
thread to shut down, and then returns anyway. This can cause problems if
there are a large number of captures outstanding, as the capture job
handler might take a while to shut down, stopAll() returns, and the
capture job handler is deleted while still in use.

To address this, and a force shutdown flag and use it in unit tests to
force all active jobs to shut down immediately.

Also make sure m_use_new_k8s is initialized. In normal agent use, this
is set by reading the configuration, but the unit test has no
configuration object.