Fix more valgrind errs (#616)
* Also actually join with all threads.

Even with the protections to force capture job shutdown, it can take
longer than 10 seconds to shut down on a very busy host under
valgrind. So join with all threads before doing the shutdown.

* Properly initialize capture job's inspector.

When a sinsp_dumper is created, it uses a temporary inspector which is
the one that reads the membuf contents. That inspector is destroyed
before returning the dumper object to the capture job handler.

Make sure the dumper has a valid inspector by setting it to the "live"
inspector that handles all future events.

Additionally, ensure that every job has a dumper object by creating it
immediately instead of only when the first matching event is found from
the membuf buffer.